@@ -1,17 +1,17 @@
 name: image_picker_type
 description: A Flutter plugin for iOS and Android for picking images from the image library and crop it.
-version: 0.0.4
+version: 1.0.0
 #author:
 homepage: https://github.com/jayeshpansheriya/image_picker_type
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  image_picker: ^0.6.7+14
-  image_cropper: ^1.3.1
+  image_picker: ^0.7.4
+  image_cropper: ^1.4.0
 
 dev_dependencies:
   flutter_test:
